requirements.txt
flask==2.3.3
flask-socketio==5.3.6
python-socketio==5.8.0
eventlet==0.33.3
PyPDF2==3.0.1
python-docx==0.8.11
pytesseract==0.3.10
pillow==10.0.0
spacy==3.7.2
nltk==3.8.1
whoosh==2.7.4
requests==2.31.0
netifaces==0.11.0